+++ paradox-language-support/src/main/kotlin/com/windea/plugin/idea/paradox/library/StellarisLibraryKind.kt
@@ -1,5 +1,6 @@
 package com.windea.plugin.idea.paradox.library
 
 import com.windea.plugin.idea.paradox.*
+import com.windea.plugin.idea.paradox.model.*
 
-object StellarisLibraryKind: ParadoxLibraryKind(stellarisName)
+object StellarisLibraryKind: ParadoxLibraryKind(ParadoxGameType.Stellaris.text)