+++ paradox-language-support/src/main/kotlin/com/windea/plugin/idea/paradox/library/StellarisLibraryType.kt
@@ -1,5 +1,6 @@
 package com.windea.plugin.idea.paradox.library
 
 import com.windea.plugin.idea.paradox.*
+import com.windea.plugin.idea.paradox.model.*
 
-class StellarisLibraryType : ParadoxLibraryType(StellarisLibraryKind, stellarisIcon, stellarisName)
+class StellarisLibraryType : ParadoxLibraryType(StellarisLibraryKind, stellarisIcon, ParadoxGameType.Stellaris.name)