+++ paradox-language-support/src/main/kotlin/com/windea/plugin/idea/paradox/model/ParadoxRootType.kt
@@ -1,5 +1,9 @@
 package com.windea.plugin.idea.paradox.model
 
-enum class ParadoxRootType{
-	Stdlib,Mod
+enum class ParadoxRootType(val text:String) {
+	Stdlib("Stdlib"),
+	Mod("Mod"),
+	PdxLauncher("Paradox Launcher"), 
+	PdxOnlineAssets("Paradox Online Assets"), 
+	TweakerGuiAssets("Tweaker GUI Assets")
 }